Require a newer OpenJDK
The 8.0.121 version was being brought in as a pin from
scijava-jupyter-kernel. We don't really need that kernel in this
environment though, so we ditch that dependency in favor of demanding a
newer and less buggy OpenJDK.

This move was motivated by the fact that 8.0.121 was giving core dumps
when attempting to use ImgLyb to talk to Java in-process from Python.